+++ CHANGELOG.md
@@ -110,7 +110,7 @@ It's also resulted in these **breaking changes**, partially because I took advan
 * Removed TestSuite from the documentation; it's now for internal use only
 
 TO DO:
-* Test watchdog failures (module load, child processes) should call onTestCaseResult.
+* Test watchdog failures (child processes) should call onTestCaseResult.
 
 ## v0.11.x: Add optional 'actual' and 'expected' to assert.fail()